+++ NEWS.md
@@ -1,5 +1,11 @@
 # googledrive (development version)
 
+* `drive_auth(subject =)` is a new argument that can be used with
+  `drive_auth(path =)`, i.e. when using a service account. The `path` and
+  `subject` arguments are ultimately processed by
+  `gargle::credentials_service_account()` and support the use of a service
+  account to impersonate a normal user (#413).
+
 * All requests now route through `gargle::request_retry()` (#380).
 
 * `drive_scopes()` is a new function to access scopes relevant to the Drive API.